refactor: move buiseness logic to custom hook, fixed restore input value from local storage
